Only mark keyword as experimental when using a Swift ≥5.8 compiler
In most other parts of the code base we only use the SPI attribute for Swift ≥5.8 compilers. If we unconditionally mark a keywords as SPI we run into issues if a syntax node (which is only experimental in Swift 5.8+) uses that keyword as a default value in its initializers. The concrete case here is the usage of `.dependsOn`.

# Conflicts:
#	Sources/SwiftSyntax/generated/Keyword.swift